requirements.txt
mss==7.0.1
Pillow==9.3.0
Flask==2.2.2
numpy==1.23.4
opencv-python==4.6.0.66
tensorflow==2.10.0
fer==22.4.0
pyqtgraph==0.13.1
requests==2.28.1
PyQt5==5.15.7
PyQtChart==5.15.6